+++ packages/client/src/views/project/simple/common/ImageActions.tsx
@@ -9,6 +9,8 @@ import {
 
 import { API } from '$api';
 import { IconButton } from '$components/iconButton';
+import { PromptLoad } from '$modals/promptLoad';
+import { modalStore } from '$stores/ModalStore';
 import { IS_ELECTRON } from '$utils/config';
 import { useSimpleProject } from '../../context';
 
@@ -58,7 +60,9 @@ export const ImageActions: React.FC<Props> = ({ type, file }) => {
           });
           const settings = data.metadata as any;
           if (settings) {
-            project.setSettings(settings);
+            modalStore.show(
+              <PromptLoad project={project} loadedSettings={settings} />,
+            );
           }
         }}
       >